old dataset.py && processor.py deprecated, and new dataset processor works
